feat: split search bar and results for better loading experience
