+++ apps/web/src/app/docs/DocumentsResults.tsx
@@ -1,12 +1,7 @@
 "use client";
 
-import { useState, useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { useDebouncedCallback } from "use-debounce";
-
 import Link from "next/link";
 import { PageLayout } from "@/components/PageLayout";
-
 import { GradeBadge } from "@/components/GradeBadge";
 import { getValidCommentCount } from "@/shared/utils/ui/commentUtils";
 import {
@@ -15,12 +10,9 @@ import {
 } from "@/shared/utils/ui/documentUtils";
 import {
   ChatBubbleLeftIcon,
-  MagnifyingGlassIcon,
   Squares2X2Icon,
   TableCellsIcon,
 } from "@heroicons/react/24/outline";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import {
@@ -93,60 +85,19 @@ interface DocumentVersionWithIncludes {
   };
 }
 
-export default function DocumentsClient({
-  documents,
-  searchQuery: initialSearchQuery,
-  totalCount,
-  hasSearched,
-  showNewButton = true,
-}: {
+interface DocumentsResultsProps {
   documents: DocumentVersionWithIncludes[];
   searchQuery: string;
   totalCount: number;
   hasSearched: boolean;
-  showNewButton?: boolean;
-}) {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const [searchQuery, setSearchQuery] = useState(initialSearchQuery);
+}
 
-  // Update local state when prop changes (e.g., from SSR)
-  useEffect(() => {
-    setSearchQuery(initialSearchQuery);
-  }, [initialSearchQuery]);
-
-  // Debounced search function that updates URL
-  const debouncedSearch = useDebouncedCallback((query: string) => {
-    const params = new URLSearchParams(searchParams.toString());
-
-    if (query.trim()) {
-      params.set("q", query.trim());
-    } else {
-      params.delete("q");
-    }
-
-    // Reset to page 1 when searching
-    params.delete("page");
-
-    const newUrl = params.toString() ? `?${params.toString()}` : "";
-    router.push(`/docs${newUrl}`);
-  }, 300);
-
-  // Handle search input change
-  const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    setSearchQuery(value);
-    debouncedSearch(value);
-  };
-
-  // Handle search input keydown
-  const handleSearchKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      debouncedSearch.flush(); // Immediately execute the search
-    }
-  };
-
+export default function DocumentsResults({
+  documents,
+  searchQuery,
+  totalCount,
+  hasSearched,
+}: DocumentsResultsProps) {
   // Get unique evaluator names from all documents
   const evaluators = Array.from(
     new Set(
@@ -163,34 +114,6 @@ export default function DocumentsClient({
 
   return (
     <PageLayout>
-      {/* Search and View Toggle - Keep in container */}
-      <div className="mx-auto max-w-7xl py-6 sm:px-6 lg:px-8">
-        <div className="px-4 py-6 sm:px-0">
-          <div className="mb-8 flex flex-col items-center gap-4">
-            <div className="flex w-full max-w-2xl justify-between">
-              <div className="relative flex-grow">
-                <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
-                  <MagnifyingGlassIcon className="h-5 w-5 text-gray-400" />
-                </div>
-                <Input
-                  type="text"
-                  value={searchQuery}
-                  onChange={handleSearchChange}
-                  onKeyDown={handleSearchKeyDown}
-                  placeholder="Search by title or agent name..."
-                  className="pl-10"
-                />
-              </div>
-              {showNewButton && (
-                <Link href="/docs/new">
-                  <Button className="ml-4">New Document</Button>
-                </Link>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-
       {/* Status message */}
       {!hasSearched && (
         <div className="py-2 text-center text-sm text-gray-600">
@@ -201,7 +124,7 @@ export default function DocumentsClient({
 
       {hasSearched && (
         <div className="py-2 text-center text-sm text-gray-600">
-          Found {totalCount} documents matching "{initialSearchQuery}"
+          Found {totalCount} documents matching "{searchQuery}"
         </div>
       )}
