+++ apps/web/src/app/docs/SearchBar.tsx
@@ -1,0 +1,89 @@
+"use client";
+
+import { useState, useEffect } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
+import { useDebouncedCallback } from "use-debounce";
+import Link from "next/link";
+import { MagnifyingGlassIcon } from "@heroicons/react/24/outline";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+
+interface SearchBarProps {
+  searchQuery: string;
+  showNewButton?: boolean;
+}
+
+export default function SearchBar({
+  searchQuery: initialSearchQuery,
+  showNewButton = true,
+}: SearchBarProps) {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const [searchQuery, setSearchQuery] = useState(initialSearchQuery);
+
+  // Update local state when prop changes (e.g., from SSR)
+  useEffect(() => {
+    setSearchQuery(initialSearchQuery);
+  }, [initialSearchQuery]);
+
+  // Debounced search function that updates URL
+  const debouncedSearch = useDebouncedCallback((query: string) => {
+    const params = new URLSearchParams(searchParams.toString());
+
+    if (query.trim()) {
+      params.set("search", query.trim());
+    } else {
+      params.delete("search");
+    }
+
+    // Reset to page 1 when searching
+    params.delete("page");
+
+    const newUrl = params.toString() ? `?${params.toString()}` : "";
+    router.push(`/docs${newUrl}`);
+  }, 300);
+
+  // Handle search input change
+  const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const value = e.target.value;
+    setSearchQuery(value);
+    debouncedSearch(value);
+  };
+
+  // Handle search input keydown
+  const handleSearchKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    if (e.key === "Enter") {
+      e.preventDefault();
+      debouncedSearch.flush(); // Immediately execute the search
+    }
+  };
+
+  return (
+    <div className="mx-auto max-w-7xl py-6 sm:px-6 lg:px-8">
+      <div className="px-4 py-6 sm:px-0">
+        <div className="mb-8 flex flex-col items-center gap-4">
+          <div className="flex w-full max-w-2xl justify-between">
+            <div className="relative flex-grow">
+              <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+                <MagnifyingGlassIcon className="h-5 w-5 text-gray-400" />
+              </div>
+              <Input
+                type="text"
+                value={searchQuery}
+                onChange={handleSearchChange}
+                onKeyDown={handleSearchKeyDown}
+                placeholder="Search by title or agent name..."
+                className="pl-10"
+              />
+            </div>
+            {showNewButton && (
+              <Link href="/docs/new">
+                <Button className="ml-4">New Document</Button>
+              </Link>
+            )}
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
