+++ apps/web/src/app/docs/loading.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+import DefaultLoading from "@/components/global-loading";
+
+export default DefaultLoading;
